Selection for wide glyphs (#905)
