+++ src/cascadia/TerminalCore/terminalcore-common.vcxproj
@@ -5,6 +5,7 @@
     <ClCompile Include="..\TerminalDispatch.cpp" />
     <ClCompile Include="..\TerminalDispatchGraphics.cpp" />
     <ClCompile Include="..\TerminalRenderData.cpp" />
+    <ClCompile Include="..\TerminalSelection.cpp" />
     <ClCompile Include="..\TerminalApi.cpp" />
     <ClCompile Include="..\Terminal.cpp" />
     <ClCompile Include="..\pch.cpp">
